+++ Cargo.toml
@@ -143,6 +143,7 @@ default = [
   "bevy_pbr",
   "bevy_picking",
   "bevy_render",
+  "bevy_render_2d",
   "bevy_scene",
   "bevy_sprite",
   "bevy_sprite_picking_backend",
@@ -242,6 +243,9 @@ bevy_picking = ["bevy_internal/bevy_picking"]
 # Provides rendering functionality
 bevy_render = ["bevy_internal/bevy_render", "bevy_color"]
 
+# Provides functionality for rendering in 2d
+bevy_render_2d = ["bevy_internal/bevy_render_2d"]
+
 # Provides scene functionality
 bevy_scene = ["bevy_internal/bevy_scene", "bevy_asset"]
 
@@ -249,6 +253,7 @@ bevy_scene = ["bevy_internal/bevy_scene", "bevy_asset"]
 bevy_sprite = [
   "bevy_internal/bevy_sprite",
   "bevy_render",
+  "bevy_render_2d",
   "bevy_core_pipeline",
   "bevy_color",
   "bevy_anti_aliasing",
