+++ crates/bevy_gizmos/Cargo.toml
@@ -16,7 +16,7 @@ bevy_render = ["dep:bevy_render", "bevy_core_pipeline"]
 [dependencies]
 # Bevy
 bevy_pbr = { path = "../bevy_pbr", version = "0.16.0-dev", optional = true }
-bevy_sprite = { path = "../bevy_sprite", version = "0.16.0-dev", optional = true }
+bevy_render_2d = { path = "../bevy_render_2d", version = "0.16.0-dev", optional = true }
 bevy_app = { path = "../bevy_app", version = "0.16.0-dev" }
 bevy_color = { path = "../bevy_color", version = "0.16.0-dev" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.16.0-dev" }
