+++ crates/bevy_internal/src/default_plugins.rs
@@ -40,6 +40,8 @@ plugin_group! {
         bevy_render::pipelined_rendering:::PipelinedRenderingPlugin,
         #[cfg(feature = "bevy_core_pipeline")]
         bevy_core_pipeline:::CorePipelinePlugin,
+        #[cfg(feature = "bevy_render_2d")]
+        bevy_render_2d::mesh_pipeline:::Mesh2dRenderPlugin,
         #[cfg(feature = "bevy_anti_aliasing")]
         bevy_anti_aliasing:::AntiAliasingPlugin,
         #[cfg(feature = "bevy_sprite")]
