+++ crates/bevy_internal/src/lib.rs
@@ -60,6 +60,8 @@ pub use bevy_reflect as reflect;
 pub use bevy_remote as remote;
 #[cfg(feature = "bevy_render")]
 pub use bevy_render as render;
+#[cfg(feature = "bevy_render_2d")]
+pub use bevy_render_2d as render_2d;
 #[cfg(feature = "bevy_scene")]
 pub use bevy_scene as scene;
 #[cfg(feature = "bevy_sprite")]
