+++ crates/bevy_render_2d/Cargo.toml
@@ -1,45 +1,30 @@
 [package]
-name = "bevy_sprite"
+name = "bevy_render_2d"
 version = "0.16.0-dev"
 edition = "2024"
-description = "Provides sprite functionality for Bevy Engine"
+description = "Provides functionality for rendering in 2d for Bevy Engine"
 homepage = "https://bevyengine.org"
 repository = "https://github.com/bevyengine/bevy"
 license = "MIT OR Apache-2.0"
 keywords = ["bevy"]
 
 [features]
-bevy_sprite_picking_backend = ["bevy_picking", "bevy_window"]
-webgl = []
-webgpu = []
 
 [dependencies]
-# bevy
 bevy_app = { path = "../bevy_app", version = "0.16.0-dev" }
 bevy_asset = { path = "../bevy_asset", version = "0.16.0-dev" }
-bevy_color = { path = "../bevy_color", version = "0.16.0-dev" }
 bevy_core_pipeline = { path = "../bevy_core_pipeline", version = "0.16.0-dev" }
+bevy_derive = { path = "../bevy_derive", version = "0.16.0-dev" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.16.0-dev" }
 bevy_image = { path = "../bevy_image", version = "0.16.0-dev" }
 bevy_math = { path = "../bevy_math", version = "0.16.0-dev" }
-bevy_picking = { path = "../bevy_picking", version = "0.16.0-dev", optional = true }
+bevy_platform_support = { path = "../bevy_platform_support", version = "0.16.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.16.0-dev" }
 bevy_render = { path = "../bevy_render", version = "0.16.0-dev" }
-bevy_render_2d = { path = "../bevy_render_2d", version = "0.16.0-dev" }
 bevy_transform = { path = "../bevy_transform", version = "0.16.0-dev" }
-bevy_utils = { path = "../bevy_utils", version = "0.16.0-dev" }
-bevy_window = { path = "../bevy_window", version = "0.16.0-dev", optional = true }
-bevy_derive = { path = "../bevy_derive", version = "0.16.0-dev" }
-bevy_platform_support = { path = "../bevy_platform_support", version = "0.16.0-dev", default-features = false, features = [
-  "std",
-] }
 
-# other
-bytemuck = { version = "1", features = ["derive", "must_cast"] }
-fixedbitset = "0.5"
-derive_more = { version = "1", default-features = false, features = ["from"] }
 bitflags = "2.3"
-radsort = "0.1"
+derive_more = { version = "1", default-features = false, features = ["from"] }
 nonmax = "0.5"
 tracing = { version = "0.1", default-features = false, features = ["std"] }
