+++ crates/bevy_render_2d/src/mesh_pipeline/shaders/mesh2d_view_bindings.wgsl
