+++ crates/bevy_sprite/src/lib.rs
@@ -30,7 +30,7 @@ pub mod prelude {
     pub use crate::{
         sprite::{Sprite, SpriteImageMode},
         texture_slice::{BorderRect, SliceScaleMode, TextureSlice, TextureSlicer},
-        ColorMaterial, MeshMaterial2d, ScalingMode,
+        ColorMaterial, ScalingMode,
     };
 }
 
@@ -96,7 +96,7 @@ impl Plugin for SpritePlugin {
             .register_type::<TextureSlicer>()
             .register_type::<Anchor>()
             .register_type::<Mesh2d>()
-            .add_plugins((Mesh2dRenderPlugin, ColorMaterialPlugin))
+            .add_plugins(ColorMaterialPlugin)
             .add_systems(
                 PostUpdate,
                 (
@@ -132,7 +132,7 @@ impl Plugin for SpritePlugin {
                     (
                         queue_sprites
                             .in_set(RenderSet::Queue)
-                            .ambiguous_with(queue_material2d_meshes::<ColorMaterial>),
+                            .before(RenderSet::QueueMeshes),
                         prepare_sprite_image_bind_groups.in_set(RenderSet::PrepareBindGroups),
                         prepare_sprite_view_bind_groups.in_set(RenderSet::PrepareBindGroups),
                         sort_binned_render_phase::<Opaque2d>.in_set(RenderSet::PhaseSort),
