+++ crates/bevy_sprite/src/mesh2d/color_material.rs
@@ -1,4 +1,3 @@
-use crate::{AlphaMode2d, Material2d, Material2dPlugin};
 use bevy_app::{App, Plugin};
 use bevy_asset::{load_internal_asset, weak_handle, Asset, AssetApp, Assets, Handle};
 use bevy_color::{Alpha, Color, ColorToComponents, LinearRgba};
@@ -6,6 +5,10 @@ use bevy_image::Image;
 use bevy_math::{Affine2, Mat3, Vec4};
 use bevy_reflect::prelude::*;
 use bevy_render::{render_asset::RenderAssets, render_resource::*, texture::GpuImage};
+use bevy_render_2d::{
+    material::Material2dPlugin,
+    prelude::{AlphaMode2d, Material2d},
+};
 
 pub const COLOR_MATERIAL_SHADER_HANDLE: Handle<Shader> =
     weak_handle!("92e0e6e9-ed0b-4db3-89ab-5f65d3678250");
