+++ crates/bevy_sprite/src/mesh2d/mod.rs
@@ -1,9 +1,5 @@
 mod color_material;
-mod material;
-mod mesh;
 mod wireframe2d;
 
 pub use color_material::*;
-pub use material::*;
-pub use mesh::*;
 pub use wireframe2d::*;
