+++ examples/shader/shader_material_2d.rs
@@ -4,7 +4,7 @@ use bevy::{
     prelude::*,
     reflect::TypePath,
     render::render_resource::{AsBindGroup, ShaderRef},
-    sprite::{AlphaMode2d, Material2d, Material2dPlugin},
+    render_2d::material::Material2dPlugin,
 };
 
 /// This example uses a shader source file from the assets subdirectory
